adjust build configuration to point to core-customize
