+++ .github/workflows/dependency-check.yml
@@ -1,18 +1,17 @@
-name: Backwards Compatibility
+name: Dependency Check
 
 on:
   push:
-    branches: [ main ]
+    branches: [ develop ]
   pull_request:
-    branches: [ main ]
+    branches: [ develop ]
 
 jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
-      fail-fast: false
       matrix:
-        version: ['2105.5', '2011.15', '2005.21']
+        extension: ['core']
     steps:
     - uses: actions/checkout@v2
     - name: Set up JDK 11
@@ -26,11 +25,8 @@ jobs:
     - name: Set up cache for SAP artifacts
       uses: actions/cache@v2
       with:
-        key: sap-artifacts-cache-${{ hashFiles('manifest.json') }}-${{ matrix.version }}
+        key: sap-artifacts-cache-${{ hashFiles('manifest.json') }}
         path: dependencies
-    - name: Configure SAP version
-      run: |
-        sed -i 's/\"commerceSuiteVersion\": \"[^\"]*\"/\"commerceSuiteVersion\": \"${{ matrix.version }}\"/' core-customize/manifest.json
     - name: SAP download config
       env:
         USER_CONFIG: ${{ secrets.GRADLE_USER_CONFIG }}
@@ -47,21 +43,11 @@ jobs:
         echo "HYBRIS_BIN_DIR=$GITHUB_WORKSPACE/core-customize/hybris/bin" >> $GITHUB_ENV
         echo "HYBRIS_CONF_DIR=$GITHUB_WORKSPACE/core-customize/hybris/config" >> $GITHUB_ENV
         echo "HYBRIS_LOG_DIR=$GITHUB_WORKSPACE/core-customize/hybris/log" >> $GITHUB_ENV
+    - name: Configure extension
+      run: |
+        cp -f ci/config/localextensions-template.xml $HYBRIS_CONF_DIR/localextensions.xml
+        sed -i 's/EXTENSIONNAME/${{ matrix.extension }}/' $HYBRIS_CONF_DIR/localextensions.xml
     - name: Bootstrap platform
       run: ./gradlew setupLocalDevelopment
     - name: Build platform
       run: ./gradlew yall
-    - name: Run unit tests
-      run: |
-        cat $GITHUB_WORKSPACE/ci/config/testing-unit.properties > $HYBRIS_CONF_DIR/local-config/99-local.properties
-        ./gradlew yunittests
-        
-        # Target never fails, therefore, check for test errors and fail if errors were found
-        for file in $HYBRIS_LOG_DIR/junit/test-results/unit/*.xml; do if [ "$(grep -cE '<(error|failure)' "${file}")" -gt 0 ]; then exit 1; fi; done
-    - name: Run integration tests
-      run: |
-        cat $GITHUB_WORKSPACE/ci/config/testing-integration.properties > $HYBRIS_CONF_DIR/local-config/99-local.properties
-        ./gradlew yintegrationtests
-        
-        # Target never fails, therefore, check for test errors and fail if errors were found
-        for file in $HYBRIS_LOG_DIR/junit/test-results/integration/*.xml; do if [ "$(grep -cE '<(error|failure)' "${file}")" -gt 0 ]; then exit 1; fi; done
