+++ ci/config/testing-integration.properties
@@ -1,4 +1,4 @@
-standalone.javaoptions=${tomcat.mem} -javaagent:${HYBRIS_BIN_DIR}/../../ci/lib/jacocoagent.jar=destfile=${HYBRIS_LOG_DIR}/junit/jacoco-integration.exec
+standalone.javaoptions=${tomcat.mem} -javaagent:${HYBRIS_BIN_DIR}/../../../ci/lib/jacocoagent.jar=destfile=${HYBRIS_LOG_DIR}/junit/jacoco-integration.exec
 installed.tenants=junit
 
 testclasses.extensions=
