+++ ci/config/testing-unit.properties
@@ -1,4 +1,4 @@
-standalone.javaoptions=${tomcat.mem} -javaagent:${HYBRIS_BIN_DIR}/../../ci/lib/jacocoagent.jar=destfile=${HYBRIS_LOG_DIR}/junit/jacoco-unit.exec
+standalone.javaoptions=${tomcat.mem} -javaagent:${HYBRIS_BIN_DIR}/../../../ci/lib/jacocoagent.jar=destfile=${HYBRIS_LOG_DIR}/junit/jacoco-unit.exec
 installed.tenants=
 
 testclasses.extensions=
