+++ core-customize/manifest.json
